refactor(charts): extract downsampleCompactLineChartData to its own module
